aws/lambda/github-webhook-sql-sync/requirements.txt
# To install the lambda layers:
#
# pip install --target ./python -r requirements.txt
# zip -r layer.zip ./python
#
# Then upload the layer via aws console or aws cli
# Update the lambda's layer reference to the newly uploaded layer

aioboto3==8.3.0
aiohttp==3.6.2
requests==2.24.0
sqlalchemy==1.4.22
pymysql==0.9.2
python-dateutil==2.8.1